fix filenames and update styles
renamed the files to match the existing filename convention. the styles are far from production ready but they should give a good sense of how we can edit the look going forward. there was a bit of a challenge with the tooltip button that required some clever uss targeting but was able to styliize it.