remove display tooltip when elided and update component uxml
As far as I know display tooltip when elided does nothing. Also message box will need some extra careful work as the Query function treats dashes and underscores as the same character so that makes it coupled with the terrarium screen uxml. Therefore the message box and the terrarium screen will have to be fixed together.