initial pass for tower detail uxml rewrite
This changes the uxml structure of the panel. The data is not hooked up properly yet and none of the buttons work. This also required a change to the styles library for buttons. This is expected since the shared styles included size and padding properties which needed to be extracted.