proof of concept for tower icon, stats, and upgrade buttons
This change gets us closer to the UI layout that we will need for the updated tower-detail panel. Many of the values are hardcoded and only the first upgrade level is visible. The stats are still not filled in. Also the upgrade mechanics need to be implemented (i.e. the button upgrade popup).